Fixed Auto-Flap Alignment, Limelight Transform, Shooter Pitch PIDF
Fixed the auto-alignment for the flaps being backwards due to left & right being from the wrong frame of reference.
Added button to swap between field-centric driving and robot-centric.
Updated shooter angle constants & added an amp state for the flaps.
Updated pose subsystem to have the new transform3d for the limelight position. When tested it gave accurate results.
Swapped the shooter's angle PID to a WPILIB PID controller instead of the sparkmax controller due to the sparkmax not going to the correct position.
Fixed flap limit switches being backwards in the periodic check.
Added a button to the smart dashboard to test the robot driving to a scoring position.
Added dashboard values to test proper angle for intaking/scoring.